requirements.txt
packaging
pillow
numpy
tensorflow
viam-sdk==0.44.0

typing-extensions
